Refactor LineColorEffects on iOS: Reuse code
